# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OpportuneX API
  description: |
    OpportuneX is an AI-powered platform that helps students from Tier 2 and Tier 3 cities in India discover and access hackathons, internships, and workshops through natural language interactions.
    
    ## Features
    - **Smart Search**: Natural language search for opportunities with advanced filtering
    - **Voice Search**: Multi-language voice commands (English and Hindi)
    - **AI Instructor**: Personalized preparation roadmaps and guidance
    - **Smart Notifications**: Personalized opportunity alerts via email, SMS, and in-app
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - General API endpoints: 100 requests per 15 minutes
    - Search endpoints: 50 requests per 15 minutes
    - Authentication endpoints: 10 requests per 15 minutes
    
    ## Error Handling
    The API uses standard HTTP status codes and returns error details in JSON format:
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Invalid request parameters",
        "details": ["Field 'query' is required"]
      }
    }
    ```
  version: 1.0.0
  contact:
    name: OpportuneX API Support
    email: api-support@opportunex.com
    url: https://opportunex.com/support
  license:
    name: MIT
...
servers:
...
paths:
...
  /search/suggestions:
    get:
      tags:
        - Search
      summary: Get search suggestions
      description: Get autocomplete suggestions for search queries
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Partial search query
          example: "AI hack"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 5
          description: Maximum number of suggestions
      responses:
        '200':
          description: Search suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    items:
                      type: string
                    example: ["AI hackathons", "AI internships", "AI workshops"]

  # Voice Search Endpoints
...
  /ai/roadmap:
    post:
      tags:
        - AI Instructor
      summary: Generate preparation roadmap
      description: |
        Generate a personalized preparation roadmap for a specific opportunity
        based on user's current skills and the opportunity requirements.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoadmapRequest'
            example:
              opportunityId: "opp_123456"
              targetDate: "2024-03-15T00:00:00Z"
              focusAreas: ["technical_skills", "project_experience"]
      responses:
        '200':
          description: Roadmap generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoadmapResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Opportunity not found
  /ai/roadmap/{roadmapId}:
    get:
      tags:
        - AI Instructor
      summary: Get roadmap details
      description: Retrieve details of a specific roadmap
      security:
        - bearerAuth: []
      parameters:
        - name: roadmapId
          in: path
          required: true
          schema:
            type: string
          description: Roadmap ID
      responses:
        '200':
          description: Roadmap details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoadmapResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Roadmap not found

    put:
      tags:
        - AI Instructor
      summary: Update roadmap progress
      description: Update progress on roadmap tasks and milestones
      security:
        - bearerAuth: []
      parameters:
        - name: roadmapId
          in: path
          required: true
          schema:
            type: string
          description: Roadmap ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completedTasks:
                  type: array
                  items:
                    type: string
                  description: Array of completed task IDs
                progress:
                  type: number
                  minimum: 0
                  maximum: 100
                  description: Overall progress percentage
      responses:
        '200':
          description: Progress updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoadmapResponse'

  # Opportunities Endpoints
  /opportunities:
    get:
      tags:
        - Opportunities
      summary: Get opportunities list
      description: Retrieve a paginated list of opportunities with optional filtering
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: ["hackathon", "internship", "workshop"]
          description: Filter by opportunity type
        - name: skills
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
          description: Filter by required skills
        - name: location
          in: query
          schema:
            type: string
          description: Filter by location
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          description: Number of items per page
      responses:
        '200':
          description: Opportunities list
          content:
            application/json:
              schema:
                type: object
                properties:
                  opportunities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Opportunity'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  totalCount:
                    type: integer
  /opportunities/{opportunityId}:
    get:
      tags:
        - Opportunities
      summary: Get opportunity details
      description: Retrieve detailed information about a specific opportunity
      security:
        - bearerAuth: []
      parameters:
        - name: opportunityId
          in: path
          required: true
          schema:
            type: string
          description: Opportunity ID
      responses:
        '200':
          description: Opportunity details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Opportunity not found

  # Notifications Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RoadmapRequest:
      type: object
      required:
        - opportunityId
      properties:
        opportunityId:
          type: string
          description: ID of the target opportunity
        targetDate:
          type: string
          format: date-time
          description: Target completion date (optional)
        focusAreas:
          type: array
          items:
            type: string
            enum: ["technical_skills", "project_experience", "interview_prep", "domain_knowledge"]
          description: Areas to focus on in the roadmap
    RoadmapResponse:
      type: object
      properties:
        id:
          type: string
          description: Roadmap ID
        opportunityId:
          type: string
          description: Associated opportunity ID
        phases:
          type: array
          items:
            $ref: '#/components/schemas/RoadmapPhase'
        estimatedDuration:
          type: integer
          description: Estimated completion time in days
        personalizedTips:
          type: array
          items:
            type: string
          description: Personalized tips and recommendations
        progress:
          type: number
          minimum: 0
          maximum: 100
          description: Current progress percentage
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RoadmapPhase:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
          description: Phase duration in days
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        resources:
          type: array
          items:
            $ref: '#/components/schemas/Resource'
        prerequisites:
          type: array
          items:
            type: string
          description: Prerequisites for this phase
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        estimatedHours:
          type: integer
          description: Estimated time to complete in hours
        priority:
          type: string
          enum: ["high", "medium", "low"]
        type:
          type: string
          enum: ["learning", "practice", "project", "assessment"]
        completed:
          type: boolean
          description: Whether the task is completed
    Resource:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: ["article", "video", "course", "book", "practice"]
        url:
          type: string
          format: uri
        duration:
          type: integer
          description: Resource duration in minutes (optional)
        difficulty:
          type: string
          enum: ["beginner", "intermediate", "advanced"]
        free:
          type: boolean
          description: Whether the resource is free

    # Opportunity Schemas
    Opportunity:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: ["hackathon", "internship", "workshop"]
        organizer:
          $ref: '#/components/schemas/Organizer'
        requirements:
          $ref: '#/components/schemas/Requirements'
        details:
          $ref: '#/components/schemas/OpportunityDetails'
        timeline:
          $ref: '#/components/schemas/Timeline'
        externalUrl:
          type: string
          format: uri
          description: Link to original opportunity posting
        sourceId:
          type: string
          description: ID of the data source
        tags:
          type: array
          items:
            type: string
          description: Opportunity tags
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isActive:
          type: boolean
          description: Whether the opportunity is currently active
    Organizer:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: ["corporate", "startup", "government", "academic"]
        logo:
          type: string
          format: uri
          description: URL to organizer logo
    Requirements:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
          description: Required technical skills
        experience:
          type: string
          description: Required experience level
        education:
          type: string
          description: Educational requirements
        eligibility:
          type: array
          items:
            type: string
          description: Eligibility criteria
    OpportunityDetails:
      type: object
      properties:
        mode:
          type: string
          enum: ["online", "offline", "hybrid"]
        location:
          type: string
          description: Event location (for offline/hybrid events)
        duration:
          type: string
          description: Event duration
        stipend:
          type: string
          description: Stipend or compensation details
        prizes:
          type: array
          items:
            type: string
          description: Prize information
    Timeline:
      type: object
      properties:
        applicationDeadline:
          type: string
          format: date-time
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time

    # Notification Schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          default: 1
        limit:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
        totalPages:
          type: integer
        totalCount:
          type: integer
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: string
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid request parameters"
              details: ["Field 'query' is required"]
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or expired token"
...